+++ .github/workflows/publish.yml
@@ -58,7 +58,6 @@ jobs:
           sudo apt install libasound2-dev
 
       - name: Build
-        working-directory: simple_audio_flutter
         run: python ./plugin_tool.py -b ${{ matrix.target }}
       - name: Upload Artifacts
         uses: actions/upload-artifact@v3.1.2
