+++ simple_audio_flutter/rust/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "simple_audio"
+name = "simple_audio_flutter"
 version = "0.1.0"
 edition = "2021"
